fix qg export and infer error in new paddle version

+++ examples/question_generation/unimo-text/export_model.py
@@ -65,14 +65,14 @@ def do_predict(args):
         unimo_text,
         input_spec=[
             # input_ids
-            paddle.static.InputSpec(shape=[None, None], dtype="int64"),
+            paddle.static.InputSpec(shape=[None, None], dtype="int32"),
             # token_type_ids
-            paddle.static.InputSpec(shape=[None, None], dtype="int64"),
+            paddle.static.InputSpec(shape=[None, None], dtype="int32"),
             # attention_mask
             paddle.static.InputSpec(shape=[None, 1, None, None],
-                                    dtype="float64"),
+                                    dtype="float32"),
             # seq_len
-            paddle.static.InputSpec(shape=[None], dtype="int64"),
+            paddle.static.InputSpec(shape=[None], dtype="int32"),
             args.max_dec_len,
             args.min_dec_len,
             args.topk,
